Arthur added Tips and initial view for goals.
I created an extendable view list for displaying tips, and created a
view to display the progress bar that represents a user's progress
towards a goal. I'm unable to show more than one at the moment. I
created 16 local tips to test the extendable view list. All of the
children contain links that will take the user to an article.